BetterChatting
- It's possible to talk directly with the bot, in private message, without refering to him - In #general channel or other open channels, it's necessary to reffer !
