+++ app/botfunc.py
@@ -1,13 +1,7 @@
-import requests
-import json
-import os
-import pprint
-from random import randint
-from dotenv import load_dotenv
-load_dotenv(".env")
 
+def bot_process_message(command):
+    response = 'I cant understand!'
+    if command.startswith('hi'):
+        response = 'hi friend'
 
-headers = {"User-agent": "curl/7.43.0", "Accept": "application/json", "user_key":os.environ.get("ZOMATO_USER_KEY")}
-
-
-
+    return response
